requirements.txt
sentence-transformers
langchain-community
langchain-pinecone
pinecone-client
python-dotenv
GitPython
langchain
streamlit
pygithub
tiktoken
requests
openai
flask
